Hw4. Все работает, исправила ошибки и картинки получаю. Сделала только свайп и как на уроке анимацию из-за этого.

+++ app/src/main/java/com/example/pictureoftheday/model/entity/DatePictureEarthGSON.kt
@@ -3,7 +3,7 @@ import com.google.gson.annotations.Expose
 
 data class DatePictureEarthGSON (
 
-	@Expose val identifier : Int,
+	@Expose val identifier : Long,
 	@Expose val caption : String,
 	@Expose val image : String,
 	@Expose val version : Int,
